requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
Pillow==10.2.0
requests==2.31.0
